Random Add Questions

+++ ChallengeCreator/app/Http/Services/QuestionService.php
@@ -191,7 +191,10 @@ class QuestionService implements BaseCrudServiceInterface
      */
     public function delete($keyOrModel): bool
     {
-        return Question::where("id",$keyOrModel)->delete();
+        if (is_int($keyOrModel)) {
+            return Question::where("id",$keyOrModel)->delete();
+        }
+        return $keyOrModel->delete();
     }
 
     /**
